requirements.txt
pynetworktables==2021.0.0
playsound==1.2.2